+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Date format: (YYYY-MM-DD)
 ## v2.23.5 (TBA)
 ### Supported MC versions: 1.21.5, 1.21.4, 1.21.3, 1.21.1, 1.21, 1.20.6
 
+* Fix: Mark shopkeepers as dirty when a property value is found missing and the default value is being used.
+  * As before, we log a warning and then fall back to the default value for missing property values. This is for example often the case when new properties are added.
+  * However, previously, we would not mark the shopkeeper as dirty in this case since the default value also equals the initial property value. One consequence of this was that the warning about missing property values would be repeated over and over during subsequent plugin reloads.
 
 ## v2.23.4 (2025-04-08)
 ### Supported MC versions: 1.21.5, 1.21.4, 1.21.3, 1.21.1, 1.21, 1.20.6